+++ OneScreen/Scan/SensorReadingParser.m
@@ -27,15 +27,17 @@ static int const kSerialNumberValueOffset = 12;
 
 - (NSDictionary*)parseData:(NSData *)manufactureData withOffset:(NSInteger)offset {
 
-    int batteryLevelValueOffset = kBatteryLevelValueOffset + offset;
-    int rhValueOffset = kRHValueOffset + offset;
-    int rhAmbientValueOffset = kRHAmbientValueOffset + offset;
-    int tempValueOffset = kTempValueOffset + offset;
-    int tempAmbientValueOffset = kTempAmbientValueOffset + offset;
+    int batteryLevelValueOffset = kBatteryLevelValueOffset + (int)offset;
+    int rhValueOffset = kRHValueOffset + (int)offset;
+    int rhAmbientValueOffset = kRHAmbientValueOffset + (int)offset;
+    int tempValueOffset = kTempValueOffset + (int)offset;
+    int tempAmbientValueOffset = kTempAmbientValueOffset + (int)offset;
 
     NSMutableDictionary *sensorData = [NSMutableDictionary dictionary];
     
     NSString* serialNumberString = [self serialNumberFromData:manufactureData withOffset:offset];
+    //if (serialNumberString == nil || serialNumberString.length != 12)
+    //    return nil;
     
     UInt8 batteryLevel = *(UInt8*)[[manufactureData subdataWithRange:NSMakeRange(batteryLevelValueOffset, 1)] bytes];
     
@@ -91,7 +93,7 @@ static int const kSerialNumberValueOffset = 12;
 
 - (NSString*)serialNumberFromData:(NSData*)data withOffset:(NSInteger) offset {
 
-    int serialNumberValueOffset = kSerialNumberValueOffset + offset;
+    int serialNumberValueOffset = kSerialNumberValueOffset + (int)offset;
 
     NSString *out2String = [[NSString alloc] initWithData:[data subdataWithRange:NSMakeRange(serialNumberValueOffset, 12)] encoding:NSASCIIStringEncoding];
